Don't require yard and git libraries to use the Rakefile
This makes it possible to use the Rakefile without those libraries
installed.  If I just want to fix a bug or add a feature to the
library, I shouldn't have to install unnecessary libraries to do
so.

This also has the advantage of speeding up the Rakefile, in most
cases, since it won't load the git library for most tasks.